Fix final gap detection
Previously, gaps at the end of the file were too easily detected,
thus causing messages to be ignored if they occurred at the end of
the last node.  Now, if there’s just white-space after the last node,
that will not be seen as a gap and it can again be messages about.

Closes wooorm/remark-lint#69.